+++ stream-chat-android-docs/src/main/kotlin/io/getstream/chat/docs/kotlin/ui/messages/MessageInput.kt
@@ -7,7 +7,6 @@ import android.view.ViewGroup
 import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
-import com.getstream.sdk.chat.utils.typing.DefaultTypingUpdatesBuffer
 import com.getstream.sdk.chat.viewmodel.MessageInputViewModel
 import io.getstream.chat.android.client.models.Attachment
 import io.getstream.chat.android.client.models.Message
@@ -52,18 +51,11 @@ private class MessageInputViewSnippets : Fragment() {
         messageInputView.setOnSendButtonClickListener {
             // Handle send button click
         }
+        // This needs to be commented out or the docs module build will fail
+        /*
         messageInputView.setTypingUpdatesBuffer(
-            DefaultTypingUpdatesBuffer(
-                onTypingStarted = {
-                    // Make a keystroke API call
-                    // Implement your custom action
-                },
-                onTypingStopped = {
-                    // Make a typing stopped API call
-                    // Implement your custom action
-                }
-            )
-        )
+            // Your custom implementation of TypingUpdatesBuffer
+        )*/
         messageInputView.setMaxMessageLengthHandler { messageText, messageLength, maxMessageLength, maxMessageLengthExceeded ->
             if (maxMessageLengthExceeded) {
                 // Show custom max-length error