+++ stream-chat-android-ui-common/src/main/kotlin/io/getstream/chat/android/common/composer/MessageComposerController.kt
@@ -98,7 +98,8 @@ public class MessageComposerController(
      */
     public var typingUpdatesBuffer: TypingUpdatesBuffer = DefaultTypingUpdatesBuffer(
         onTypingStarted = ::sendKeystrokeEvent,
-        onTypingStopped = ::sendStopTypingEvent
+        onTypingStopped = ::sendStopTypingEvent,
+        coroutineScope = scope
     )
 
     /**